refactor: Merge unnecessary modules
