+++ preferences-commons/src/main/java/no/sikt/nva/person/preferences/commons/PersonPreferencesTransactionConstants.java
@@ -1,4 +1,4 @@
-package no.sikt.nva.person.preferences.storage;
+package no.sikt.nva.person.preferences.commons;
 
 public final class PersonPreferencesTransactionConstants {
