+++ preferences-commons/src/main/java/no/sikt/nva/person/preferences/commons/service/PersonPreferencesService.java
@@ -13,7 +13,7 @@ import java.util.List;
 import java.util.Map;
 
 import static java.util.Objects.isNull;
-import static no.sikt.nva.person.preferences.storage.PersonPreferencesTransactionConstants.PRIMARY_PARTITION_KEY;
+import static no.sikt.nva.person.preferences.commons.PersonPreferencesTransactionConstants.PRIMARY_PARTITION_KEY;
 
 public class PersonPreferencesService {
