+++ preferences-rest/src/test/java/no/sikt/nva/person/preferences/rest/UpsertPersonPreferencesHandlerTest.java
@@ -4,7 +4,7 @@ import com.amazonaws.services.lambda.runtime.Context;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import no.sikt.nva.person.preferences.commons.model.PersonPreferences;
 import no.sikt.nva.person.preferences.commons.service.PersonPreferencesService;
-import no.sikt.nva.person.preferences.test.support.LocalPreferencesTestDatabase;
+import no.sikt.nva.person.preferences.commons.LocalPreferencesTestDatabase;
 import no.unit.nva.testutils.HandlerRequestBuilder;
 import nva.commons.apigateway.GatewayResponse;
 import nva.commons.apigateway.exceptions.NotFoundException;
