--- preferences-storage/build.gradle
@@ -1,9 +1,0 @@
-plugins {
-    id 'nva.person.preferences.java-conventions'
-}
-
-dependencies {
-    testRuntimeOnly(nvaCatalog.datafaker)
-    testRuntimeOnly(nvaCatalog.jackson.core)
-    testRuntimeOnly(nvaCatalog.jackson.databind)
-}
