--- preferences-test-support/build.gradle
@@ -1,11 +1,0 @@
-plugins {
-    id 'nva.person.preferences.java-conventions'
-}
-
-dependencies {
-    api(nvaCatalog.aws.sdk.dynamo)
-    api(nvaCatalog.junit.jupiter.api)
-    implementation(nvaCatalog.aws.dynamodb.local)
-    implementation(project(":preferences-storage"))
-    runtimeOnly(nvaCatalog.datafaker)
-}
